* Update test_case.zip. * Update VC++ project file. * Correct several problem in the output. * Fix a mistake with windows calling convention. * Now j0llyDmpr can be compiled in x64 without warnings/errors.

+++ usb.h
@@ -36,7 +36,7 @@ UCHAR GetLetterOfNewVolume(const DWORD precVal, const DWORD newVal);
 BOOL DumpAndSearchInteresstingFiles(const PUCHAR pVol, const DWORD lvl, PCONFIG pConf);
 
 /* # Filename match with patterns */
-BOOL isAnInteresstingFile(const PUCHAR file, const unsigned long long fileSize, PCONFIG pConf);
+BOOL isAnInteresstingFile(const PUCHAR file, const DWORD64 fileSize, PCONFIG pConf);
 
 /* # Initialization stuff */
 VOID initUsbStuff(const char* outpath);
